+++ dwarfeng-ftp-api/src/test/java/com/dwarfeng/ftp/api/integration/example/TelqosExample.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.ftp.api.integration.example;
 
+import com.dwarfeng.dutil.basic.io.CT;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.util.Scanner;
@@ -22,14 +23,14 @@ public class TelqosExample {
         Scanner scanner = new Scanner(System.in);
 
         // 显示欢迎信息。
-        System.out.println("开发者您好!");
-        System.out.println("这是一个示例, 用于演示 dwarfeng-ftp 的 Telqos 功能");
-        System.out.println("您可以使用 telnet 客户端工具访问本机 ${telqos.port} 端口以体验本示例的功能");
-        System.out.println("您可以使用 telqos 的 ftp 命令以操作 ftp qos 服务");
+        CT.trace("开发者您好!");
+        CT.trace("这是一个示例, 用于演示 dwarfeng-ftp 的 Telqos 功能");
+        CT.trace("您可以使用 telnet 客户端工具访问本机 ${telqos.port} 端口以体验本示例的功能");
+        CT.trace("您可以使用 telqos 的 ftp 命令以操作 ftp qos 服务");
         System.out.print("本示例的功能体验完成后，请按回车键结束示例...");
         scanner.nextLine();
 
-        System.out.println("示例演示完毕, 感谢您测试与使用!");
+        CT.trace("示例演示完毕, 感谢您测试与使用!");
 
         System.exit(0);
     }